Fix intonation. This bug was caused by the commit 7789b795 (2015-08-15).

+++ src/vtm_control_model/EventList.cpp
@@ -1015,7 +1015,7 @@ EventList::generateOutput(std::ostream& vtmParamStream)
 		}
 		if (!microFlag_) table[0] = 0.0;
 		if (driftFlag_)  table[0] += static_cast<float>(driftGenerator_.drift());
-		if (macroFlag_)  table[0] += static_cast<long>(currentValues[32]);
+		if (macroFlag_)  table[0] += static_cast<float>(currentValues[32]);
 
 		table[0] += static_cast<float>(pitchMean_);
